Moved service unit tests into tests directory
