+++ src/app/tests/services/auth.service.spec.ts
@@ -1,8 +1,8 @@
 import { TestBed } from '@angular/core/testing';
 
-import { AuthService } from '../../app/services/auth.service';
+import { AuthService } from '../../services/auth.service';
 import { AngularFireModule } from '@angular/fire';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { AngularFirestore } from '@angular/fire/firestore';
 import { AngularFireAuth } from '@angular/fire/auth';
 import { RouterTestingModule } from '@angular/router/testing';
