+++ src/app/tests/services/message.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { MessageService } from './message.service';
+import { MessageService } from '../../services/message.service';
 import { AngularFirestore } from '@angular/fire/firestore';
 
 //  test suite
